- separated BaseUrl from Crawler and Website in own file. - logger now a function from base.
